requirements.txt
Django
djangorestframework
Pillow
psycopg
pytz
django-filter

